+++ src/lib/Components/Bidding/__stories__/BidFlow.story.tsx
@@ -47,7 +47,7 @@ storiesOf("Bidding")
       <ConfirmBid
         sale_artwork={{
           _id: "saleartwork12345",
-          sale: { id: "sale-id", live_start_at: "2018-06-11T01:00:00+00:00", end_at: null },
+          sale: { id: "sale-id", live_start_at: "2018-08-13T18:00:00+00:00", end_at: null },
           artwork: { id: "artwork-id", title: "Morgan Hill (Prototype)", date: "1973", artist_names: "Lewis balts" },
           lot_label: "2",
         }}