+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 
 * Fixes a bug where the error message for the country field remains even after selecting country - yuki24
 * Adds an error modal to the bid flow - sweir27
+* The country select form should pre-populate if user have already selected a country - yuki2
 * Update graphql + use new argument to fetch increments - sweir27
 
 ### 1.5.10