+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 
 ### Master
 
+### 1.5.11
+
 * Fixes a bug where the error message for the country field remains even after selecting country - yuki24
 * Adds an error modal to the bid flow - sweir27
 * The country select form should pre-populate if user have already selected a country - yuki24